backend/requirements.txt
asgiref==3.7.2
coverage==7.4.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
